Refactor fd caching mechanism in a separate module
I will need this soon to also gather `TCP_INFO` stats.
