+++ ndt-server.go
@@ -18,8 +18,9 @@ import (
 
 	"github.com/gorilla/handlers"
 	"github.com/gorilla/websocket"
-	"github.com/m-lab/ndt-cloud/ndt7"
 	"github.com/m-lab/ndt-cloud/bbr"
+	"github.com/m-lab/ndt-cloud/fdcache"
+	"github.com/m-lab/ndt-cloud/ndt7"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 )
@@ -130,10 +131,18 @@ func (ln tcpListenerEx) Accept() (net.Conn, error) {
 	}
 	tc.SetKeepAlive(true)
 	tc.SetKeepAlivePeriod(3 * time.Minute)
+	fp, err := fdcache.TCPConnToFile(tc)
+	if err != nil {
+		tc.Close()
+		return nil, err
+	}
 	if ln.TryToEnableBBR {
-		err = bbr.EnableAndRememberFile(tc)
+		err = bbr.Enable(fp)
 		if err != nil && err != bbr.ErrNoSupport {
 			log.Printf("Cannot initialize BBR: %s", err.Error())
+			// We need to close both because fp is a dup() of the original tc.
+			fp.Close()
+			tc.Close()
 			return nil, err
 		}
 		if err == bbr.ErrNoSupport {
@@ -142,6 +151,9 @@ func (ln tcpListenerEx) Accept() (net.Conn, error) {
 			// where the operating system is different from Linux.
 		}
 	}
+	// Transfer ownership of |fp| to fdcache so that later we can retrieve
+	// it from the generic net.Conn object bound to a websocket.Conn.
+	fdcache.OwnFile(tc, fp)
 	return tc, nil
 }
